Cleanup & optimization
