+++ even-more-fish-plugin/src/main/java/com/oheers/fish/utils/ItemFactory.java
@@ -1,10 +1,8 @@
 package com.oheers.fish.utils;
 
-import com.denizenscript.denizen.objects.ItemTag;
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.FishUtils;
 import com.oheers.fish.config.messages.Message;
-import dev.lone.itemsadder.api.CustomStack;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.OfflinePlayer;
